+++ programs/clearing_house/src/state/fulfillment.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, PartialEq, Eq)]
 pub enum PerpFulfillmentMethod {
-    AMM(Option<u128>),
+    AMM(Option<u64>),
     Match,
 }
