+++ programs/clearing_house/src/validation/margin.rs
@@ -32,7 +32,7 @@ pub fn validate_margin(
     )?;
 
     validate!(
-        (margin_ratio_initial as u128) * 100 > max_spread as u128,
+        margin_ratio_initial * 100 > max_spread,
         ErrorCode::InvalidMarginRatio,
         "margin_ratio_initial must be greater than max_spread (or must lower max_spread first)"
     )?;
